Move electron start building application errors
