--- src/electron.js
@@ -1,87 +1,0 @@
-const electron = require('electron');
-// Module to control application life.
-const app = electron.app;
-// Module to create native browser window.
-const BrowserWindow = electron.BrowserWindow;
-
-const os = require('os');
-const path = require('path');
-const { ipcMain } = require( "electron" );
-
-// Keep a global reference of the window object, if you don't, the window will
-// be closed automatically when the JavaScript object is garbage collected.
-let mainWindow;
-
-// app.removeAllListeners('ready');
-
-function createWindow() {
-	BrowserWindow.addDevToolsExtension(
-		path.join(
-			os.homedir(),
-			`/AppData/Local/Google/Chrome/User Data/Default/Extensions/fmkadmapgofadopljbjfkapdkoienihi/4.4.0_0`,
-		),
-	);
-
-	// Create the browser window.
-	mainWindow = new BrowserWindow({
-		width: 1200,
-		height: 800,
-		webPreferences: { 
-			nodeIntegration: true,
-			preload: __dirname + '/preload.js' 
-		},
-	});
-
-	mainWindow.removeMenu();
-	// and load the index.html of the app.
-	mainWindow.loadURL('http://localhost:3000');
-
-	// Open the DevTools.
-	mainWindow.webContents.openDevTools();
-
-	// Emitted when the window is closed.
-	mainWindow.on('closed', function() {
-		// Dereference the window object, usually you would store windows
-		// in an array if your app supports multi windows, this is the time
-		// when you should delete the corresponding element.
-		mainWindow = null;
-	});
-}
-
-// This method will be called when Electron has finished
-// initialization and is ready to create browser windows.
-// Some APIs can only be used after this event occurs.
-app.on('ready', createWindow);
-
-// Quit when all windows are closed.
-app.on('window-all-closed', function() {
-	// On OS X it is common for applications and their menu bar
-	// to stay active until the user quits explicitly with Cmd + Q
-	if (process.platform !== 'darwin') {
-		app.quit();
-	}
-});
-
-app.on('activate', function() {
-	// On OS X it's common to re-create a window in the app when the
-	// dock icon is clicked and there are no other windows open.
-	if (mainWindow === null) {
-		createWindow();
-	}
-});
-
-// In this file you can include the rest of your app's specific main process
-// code. You can also put them in separate files and require them here.
-
-const dialog = require('electron').dialog;
-
-ipcMain.on('selectPackagePath', async event => {
-	const dialogResult = await dialog.showSaveDialog({
-			title:'This is this',
-			buttonLabel: 'Create Package',
-			message: 'Create a developers error report package',
-		})
-		const savePath = dialogResult.filePath;
-
-		event.sender.send('selectedPackagePath',savePath);
-})
