+++ src/electron/preload.js
@@ -1,0 +1,4 @@
+
+
+// @ts-ignore
+window.ipcRenderer = require('electron').ipcRenderer;
