--- src/preload.js
@@ -1,1 +1,0 @@
-window.ipcRenderer = require('electron').ipcRenderer;
